pyproject.toml
[project]
name = "py_template"
version = "0.1.0"
description = "General purpose Python project template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "linearmodels>=7.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.14.9",
    "ty>=0.0.1a35",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
